chore: 검색 페이지 비지니스로직, UI 분리
