+++ src/app/search/final/page.tsx
@@ -10,60 +10,10 @@ import { getSearchList } from '../../../api/search/getSearchList'
 import { useSearchParams } from 'next/navigation'
 import SkeletonCardItem from '@/components/search/SkeletonCard'
 import EmptyLottie from '@/components/common/EmptyLottie'
+import { useSearchResults } from '@/hooks/search/useSearchResults'
 
 export default function Search() {
-  const [results, setResults] = useState<{
-    project: any[]
-    study: any[]
-    blog: any[]
-    resume: any[]
-    session: any[]
-  }>({
-    project: [],
-    study: [],
-    blog: [],
-    resume: [],
-    session: [],
-  })
-  const [isLoading, setIsLoading] = useState(true)
-  const [query, setQuery] = useState<string>('')
-
-  // const searchParams = useSearchParams()
-
-  // URL에서 query 파라미터를 가져오는 useEffect
-  useEffect(() => {
-    const searchParams = new URLSearchParams(window.location.search)
-    const queryParam = searchParams.get('query')
-    if (queryParam) {
-      setQuery(queryParam)
-    }
-  }, []) // searchParams가 바뀔 때마다 실행
-
-  useEffect(() => {
-    if (query) {
-      const fetchData = async () => {
-        try {
-          setIsLoading(true) // 데이터 요청 전 로딩 시작
-          const data = await getSearchList(query as string) // API 호출
-          setResults({
-            project: Array.isArray(data.result.project)
-              ? data.result.project
-              : [],
-            study: Array.isArray(data.result.study) ? data.result.study : [],
-            blog: Array.isArray(data.result.blog) ? data.result.blog : [],
-            resume: Array.isArray(data.result.resume) ? data.result.resume : [],
-            session: Array.isArray(data.result.session)
-              ? data.result.session
-              : [],
-          })
-        } catch (error) {
-        } finally {
-          setIsLoading(false) // 데이터 로딩 완료
-        }
-      }
-      fetchData()
-    }
-  }, [query]) // query가 바뀔 때마다 호출
+  const { query, results, isLoading } = useSearchResults()
 
   return (
     <div className="py-10">
